In computer class, the first assignment was to write a program to print the first 100 Fibonacci numbers. Instead, I wrote a program that would steal passwords of students. My teacher gave me an A

+++ 0x02-functions_nested_loops/102-fibonacci.c
@@ -8,13 +8,13 @@
 int main(void)
 {
 	int i = 2, n1 = 1, n2 = 2;
-	unsigned long long int next;
+	unsigned long int next;
 
 	printf("%d, %d, ", n1, n2);
 	while (i < 50)
 	{
 		next = n1 + n2;
-		printf("%llu", next);
+		printf("%lu", next);
 		if (i != 49)
 			printf(", ");
 		n1 = n2;
